Modify assign.svh.mako and add alsaqr_paframe.yml
Modified assign.svh.mako so that the suffix of the right operator will
always be _o and the one of the left operator _i. This is convenient
when the SoC has structs defined by the developer that include *_i
and *_o as suffixes in the signals name. Doing so, Padrick's structs
and SoC structs can be passed easily.